LPD-72877 Move StructureService to common
